requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
Pillow==8.4.0
sqlparse==0.4.1
asgiref==3.4.1
pytz==2021.1
